+++ R/server/stages/120_reset.R
@@ -1,6 +1,7 @@
   # ---------------- Reset ----------------
   reset_pipeline_state <- function() {
     rv$results <- list(); rv$summary <- NULL; rv$rename <- NULL; rv$settings <- NULL
+    rv$assay_profiles <- assay_default_profiles(); rv$project_defaults <- assay_project_defaults_from_legacy_settings()
     rv$read_assignments <- stage2_empty_assignments(); rv$architecture <- NULL; rv$consensus_set <- stage3_empty_consensus_set(); rv$multilocus_profile <- stage4_empty_profile(); rv$project_migration_log <- ""
     rv$blast_jobs <- rv$blast_jobs[0,]; rv$blast_raw <- list(); rv$blast_ids <- data.frame(); rv$blast_hits <- data.frame()
     rv$ncbi_last_contact <- as.POSIXct(NA); rv$blast_batch_status_text <- "No batch operation has been run yet."